update TextInputLayoutUtils to target new material design version
