+++ build.gradle
@@ -17,6 +17,7 @@ buildscript {
             junit                : '4.12',
             leakcanary           : '1.6.1',
             lightweightStream    : '1.2.1',
+            materialDesign       : '1.0.0',
             mockito              : '2.18.0',
             okhttp3              : '3.11.0',
             picasso              : '2.5.2',
