+++ gto-support-design/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
     implementation "com.android.support:design:${deps.androidSupport}"
+    implementation "com.google.android.material:material:${deps.materialDesign}"
 
     implementation "com.jakewharton.timber:timber:${deps.timber}"
 }
